IOS-3987 Fix review comments
